feat: add dynamic income driver form component
