+++ src/app/components/wizard-shell/wizard-shell.component.html
@@ -115,14 +115,14 @@
 
     <!-- Income Drivers Component (Existing Store only) -->
     <div *ngIf="answers.storeType === 'existing'">
-      <app-income-drivers
+      <app-legacy-income-drivers
         [region]="answers.region"
         [handlesTaxRush]="answers.handlesTaxRush || false"
         [hasOtherIncome]="answers.hasOtherIncome || false"
         [initialData]="getIncomeDriverData()"
         (dataChange)="onIncomeDriverChange($event)"
         (calculatedValues)="onIncomeCalculatedValues($event)">
-      </app-income-drivers>
+      </app-legacy-income-drivers>
     </div>
 
     <!-- Target Performance Goals Section (New Store only) -->
